# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "DliDli 视频社区后端 API（C 端 + 管理后台）。统一响应：{ code, message, data, trace_id }，code=0 为成功。",
        "title": "DliDli API",
        "contact": {},
        "version": "0.1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/admin/admins": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理后台-RBAC"
                ],
                "summary": "管理员账号列表",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_dlidli_server_internal_pkg_response.Body"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理后台-RBAC"
                ],
                "summary": "新建管理员账号",
                "parameters": [
                    {
                        "description": "username; nickname; password; role_ids",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_module_admin.SaveAdminReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_dlidli_server_internal_pkg_response.Body"
                        }
                    }
                }
            }
        },
        "/admin/admins/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理后台-RBAC"
                ],
                "summary": "编辑管理员账号",
                "parameters": [
                    {
                        "type": "string",
                        "description": "账号ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "nickname; role_ids",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_module_admin.SaveAdminReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_dlidli_server_internal_pkg_response.Body"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理后台-RBAC"
                ],
                "summary": "删除管理员账号",
                "parameters": [
                    {
                        "type": "string",
                        "description": "账号ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_dlidli_server_internal_pkg_response.Body"
                        }
                    }
                }
            }
        },
        "/admin/admins/{id}/reset-password": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理后台-RBAC"
                ],
                "summary": "重置账号密码",
                "parameters": [
                    {
                        "type": "string",
                        "description": "账号ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "password",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_module_admin.ResetPwdReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_dlidli_server_internal_pkg_response.Body"
                        }
                    }
                }
            }
        },
        "/admin/admins/{id}/toggle": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理后台-RBAC"
                ],
                "summary": "启用/停用账号",
                "parameters": [
                    {
                        "type": "string",
                        "description": "账号ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "status: 0启用/1停用",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_dlidli_server_internal_pkg_response.Body"
                        }
                    }
                }
            }
        },
        "/admin/categories": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理后台-分区"
                ],
                "summary": "分区列表（含停用）",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_dlidli_server_internal_pkg_response.Body"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理后台-分区"
                ],
                "summary": "新建分区",
                "parameters": [
                    {
                        "description": "parent_id; name; sort; status",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_dlidli_server_internal_module_video.SaveCategoryReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_dlidli_server_internal_pkg_response.Body"
                        }
                    }
                }
            }
        },
        "/admin/categories/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理后台-分区"
                ],
                "summary": "编辑分区",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "分区ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "name; sort; status",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_dlidli_server_internal_module_video.SaveCategoryReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_dlidli_server_internal_pkg_response.Body"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理后台-分区"
                ],
                "summary": "删除分区（有子分区或稿件时禁删）",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "分区ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_dlidli_server_internal_pkg_response.Body"
                        }
                    }
                }
            }
        },
        "/admin/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理后台"
                ],
                "summary": "后台登录",
                "parameters": [
                    {
                        "description": "管理员账密",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_module_admin.LoginReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "data: {token, username, role}",
                        "schema": {
                            "$ref": "#/definitions/github_com_dlidli_server_internal_pkg_response.Body"
                        }
                    }
                }
            }
        },
        "/admin/me/permissions": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理后台-RBAC"
                ],
                "summary": "当前登录者权限码与可见菜单",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_dlidli_server_internal_pkg_response.Body"
                        }
                    }
                }
            }
        },
        "/admin/permissions": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理后台-RBAC"
                ],
                "summary": "权限点全集",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_dlidli_server_internal_pkg_response.Body"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理后台-权限点"
                ],
                "summary": "新建权限点（页面 menu / 按钮 button）",
                "parameters": [
                    {
                        "description": "code; name; type=menu|button; parent; path; icon; sort",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_module_admin.SavePermissionReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_dlidli_server_internal_pkg_response.Body"
                        }
                    }
                }
            }
        },
        "/admin/permissions/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理后台-权限点"
                ],
                "summary": "编辑权限点（code 锁定）",
                "parameters": [
                    {
                        "type": "string",
                        "description": "权限点ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "name; type; parent; path; icon; sort",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_module_admin.SavePermissionReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_dlidli_server_internal_pkg_response.Body"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理后台-权限点"
                ],
                "summary": "删除权限点（有子节点或被角色引用时禁删）",
                "parameters": [
                    {
                        "type": "string",
                        "description": "权限点ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_dlidli_server_internal_pkg_response.Body"
                        }
                    }
                }
            }
        },
        "/admin/roles": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理后台-RBAC"
                ],
                "summary": "角色列表",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_dlidli_server_internal_pkg_response.Body"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理后台-RBAC"
                ],
                "summary": "新建角色",
                "parameters": [
                    {
                        "description": "name; code; remark; perms",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_module_admin.SaveRoleReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_dlidli_server_internal_pkg_response.Body"
                        }
                    }
                }
            }
        },
        "/admin/roles/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理后台-RBAC"
                ],
                "summary": "编辑角色/分配权限",
                "parameters": [
                    {
                        "type": "string",
                        "description": "角色ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "name; remark; perms",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_module_admin.SaveRoleReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_dlidli_server_internal_pkg_response.Body"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理后台-RBAC"
                ],
                "summary": "删除角色（内置角色禁删）",
                "parameters": [
                    {
                        "type": "string",
                        "description": "角色ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_dlidli_server_internal_pkg_response.Body"
                        }
                    }
                }
            }
        },
        "/admin/sensitive-words": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理后台"
                ],
                "summary": "敏感词列表",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_dlidli_server_internal_pkg_response.Body"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理后台"
                ],
                "summary": "新增敏感词",
                "parameters": [
                    {
                        "description": "word",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_module_admin.AddWordReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_dlidli_server_internal_pkg_response.Body"
                        }
                    }
                }
            }
        },
        "/admin/sensitive-words/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理后台"
                ],
                "summary": "删除敏感词",
                "parameters": [
                    {
                        "type": "string",
                        "description": "敏感词ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_dlidli_server_internal_pkg_response.Body"
                        }
                    }
                }
            }
        },
        "/admin/users": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理后台"
                ],
                "summary": "用户查询（UID/手机号/昵称+状态）",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UID/手机号/昵称",
                        "name": "keyword",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "状态 -1全部/0正常/1禁言/2封禁",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_dlidli_server_internal_pkg_response.Body"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_dlidli_server_internal_pkg_response.Body": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                },
                "trace_id": {
                    "type": "string"
                }
            }
        },
...
        "internal_module_admin.SaveAdminReq": {
            "type": "object",
            "required": [
                "username"
            ],
            "properties": {
                "nickname": {
                    "type": "string",
                    "maxLength": 32
                },
                "password": {
                    "type": "string",
                    "maxLength": 64
                },
                "role_ids": {
                    "description": "雪花 ID 字符串化（防 JS 精度丢失）",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "username": {
                    "type": "string",
                    "maxLength": 32
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "在值中填入：Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}